Pull Request - Guardian, Sf, Sp

+++ src/main/scala/models/Quote.scala
@@ -1,9 +1,9 @@
 package models
 
 final case class Quote(timestamp: Long,
-                        open: Double,
-                        low: Double,
-                        volume: Double,
-                        high: Double,
-                        close: Double,
-                        adjclose: Double)
+                        open:  BigDecimal,
+                        low: BigDecimal,
+                        volume: BigDecimal,
+                        high: BigDecimal,
+                        close: BigDecimal,
+                        adjclose: BigDecimal)
